Traffic Sign Prediction/requirements.txt
flask
opencv-python
Pillow
keras
tensorflow
scikit-image
numpy
random2
strings